user context checking to allow  mutations

+++ src/datasources/ingredient.ts
@@ -1,5 +1,5 @@
 import { DataSource, DataSourceConfig } from 'apollo-datasource';
-import { dbIngredient, IngredientModel } from '../store';
+import { dbIngredient, dbRecipe, IngredientModel } from '../store';
 import { Context } from '..';
 
 interface CreateIngredientParams {
@@ -21,9 +21,7 @@ class Ingredient extends DataSource {
   }
 
   async getIngredients(recipeId: string): Promise<Array<IngredientModel>> {
-    return await dbIngredient.findAll({
-      where: { recipeId },
-    });
+    return await dbIngredient.findAll({ where: { recipeId } });
   }
 
   async createIngredient({
@@ -32,20 +30,36 @@ class Ingredient extends DataSource {
     recipeId,
   }: CreateIngredientParams): Promise<IngredientModel | null> {
     if (!this.context.user) return null;
+    const recipe = await dbRecipe.findOne({
+      where: { id: recipeId, ownerId: this.context.user.id },
+    });
+    if (!recipe) return null;
     return await dbIngredient.create({ text, isChecked, recipeId });
   }
 
   async updateIngredient(
     updatedIngredient: UpdateIngredientParams,
-    ingredientId: string
+    ingredientId: string,
+    recipeId: string
   ): Promise<IngredientModel | null> {
     if (!this.context.user) return null;
-    const ingredient = await dbIngredient.findOne({ where: { id: ingredientId } });
+    const recipe = await dbRecipe.findOne({
+      where: { id: recipeId, ownerId: this.context.user.id },
+    });
+    if (!recipe) return null;
+    const ingredient = await dbIngredient.findOne({
+      where: { id: ingredientId, recipeId },
+    });
     if (!ingredient) return null;
     return await ingredient.update(updatedIngredient);
   }
 
-  async deleteIngredient(ingredientId: string): Promise<boolean> {
+  async deleteIngredient(ingredientId: string, recipeId: string): Promise<boolean> {
+    if (!this.context.user) return false;
+    const recipe = await dbRecipe.findOne({
+      where: { id: recipeId, ownerId: this.context.user.id },
+    });
+    if (!recipe) return false;
     return (await dbIngredient.destroy({ where: { id: ingredientId } })) === 1;
   }
 }
